Fixxing bug in MemoryUilization() func  Linux_parser.cpp: change MemTotal to MemTotal: and MemFree to MemFree:

+++ Projects/CppND-System-Monitor/include/processor.h
@@ -7,10 +7,10 @@ class Processor {
 
   // TODO: Declare any necessary private members
  private:
-  long PrevIdle{0};
-  long PrevNonIdle{0};
-  long Idle{0};
-  long NonIdle{0};
+  float PrevIdle{0};
+  float PrevNonIdle{0};
+  float Idle{0};
+  float NonIdle{0};
 };
 
 #endif
